Criando classe de error

+++ src/storage/group/groupCreate.ts
@@ -1,12 +1,18 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
-
-import { GROUP_COLLECTION } from '@storage/storageConfig';
+import { AppError } from "@utils/appError";
 
 import { groupsGetAll } from "./groupsGetAll";
+import { GROUP_COLLECTION } from '@storage/storageConfig';
 
-export async function groupCreate(newGroup: string){
+export async function groupCreate(newGroup: string) {
   try {
     const storedGroups = await groupsGetAll();
+
+    const groupAllreadyExists = storedGroups.includes(newGroup);
+
+    if (groupAllreadyExists) {
+      throw new AppError("Já existe um grupo cadastrado com esse nome.");
+    }
 
     const storage = JSON.stringify([...storedGroups, newGroup]);
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
